Initial checkin.
First stab at a PHP web application framework which:

(1) is simple enough to be completely understood.
(2) leans heavily on "stock" PHP functionality
(3) is as fast as a non-compiled/PHP application can be
(4) has enough functionality to substantially accerate the development of a database-backed API

My greatest interest is in item (1), so the design leans toward simplicity.